+++ hata/main/commands/default/__init__.py
@@ -2,4 +2,5 @@ __all__ = ()
 
 from .help import *
 from .interpreter import *
+from .run import *
 from .version import *